+++ android/src/main/java/dev/barooni/capacitor/calendar/utils/ImplementationHelper.kt
@@ -1,0 +1,24 @@
+package dev.barooni.capacitor.calendar.utils
+
+import com.getcapacitor.JSArray
+import dev.barooni.capacitor.calendar.PluginError
+import java.util.Calendar
+
+class ImplementationHelper {
+    companion object {
+        fun getCalendarFromTimestamp(timestamp: Long?): Calendar =
+            Calendar.getInstance().apply {
+                timeInMillis = timestamp ?: System.currentTimeMillis()
+            }
+
+        fun jsArrayToComaSeparatedString(array: JSArray?): String? {
+            val list = array?.toList<Any>() ?: return null
+
+            if (!list.all { it is String }) {
+                throw PluginError.InvalidInvitees
+            }
+
+            return list.joinToString(", ")
+        }
+    }
+}