+++ src/index.ts
@@ -11,6 +11,8 @@ import type { CalendarSource } from "./schemas/interfaces/calendar-source";
 import { CalendarSourceType } from "./schemas/enums/calendar-source-type";
 import { CalendarType } from "./schemas/enums/calendar-type";
 import type { CapacitorCalendarPlugin } from "./definitions";
+import type { CreateEventWithPromptOptions } from "./sub-definitions/event-operations";
+import { EventAvailability } from "./schemas/enums/event-availability";
 import { EventSpan } from "./schemas/enums/event-span";
 import type { Reminder } from "./schemas/interfaces/reminder";
 import { ReminderRecurrenceFrequency } from "./schemas/enums/reminder-recurrence-frequency";
@@ -26,7 +28,6 @@ const CapacitorCalendar = registerPlugin<CapacitorCalendarPlugin>(
 );
 
 export * from "./definitions";
-export { CalendarPermissionScope };
 export type {
   Calendar,
   RemindersList,
@@ -34,6 +35,7 @@ export type {
   CalendarSource,
   CheckAllPermissionsResult,
   RequestAllPermissionsResult,
+  CreateEventWithPromptOptions,
 };
 export {
   CalendarChooserSelectionStyle,
@@ -44,5 +46,7 @@ export {
   EventSpan,
   CalendarSourceType,
   CalendarType,
+  CalendarPermissionScope,
+  EventAvailability,
   CapacitorCalendar,
 };